rename env to settings
