+++ src/app/settings.ts
@@ -1,6 +1,6 @@
-import env from "codeforlife/env"
+import env from "codeforlife/settings"
 
-export * from "codeforlife/env"
+export * from "codeforlife/settings"
 
 export const LINK_CFL = env.VITE_LINK_CFL
 export const LINK_GH_LOGIN = env.VITE_LINK_GH_LOGIN
