+++ src/pages/home/Home.tsx
@@ -11,7 +11,7 @@ import { GitHub as GitHubIcon } from "@mui/icons-material"
 import { Image } from "codeforlife/components"
 import { LinkButton } from "codeforlife/components/router"
 
-import { LINK_CFL, LINK_GH_LOGIN } from "../../app/env"
+import { LINK_CFL, LINK_GH_LOGIN } from "../../app/settings"
 import CflLogoImage from "../../images/cfl_logo.png"
 import { paths } from "../../routes"
 import { useLoginMutation } from "../../api/session"
